+++ criteria/parser.go
@@ -136,11 +136,10 @@ outer:
 			leftOperand := toolbox.AsString(criterion.LeftOperand)
 			if strings.HasSuffix(leftOperand, "!") {
 				criterion.LeftOperand = string(leftOperand[:len(leftOperand)-1])
-				criterion.Operator = "!"  + token.Matched
+				criterion.Operator = "!" + token.Matched
 			} else {
 				criterion.Operator = token.Matched
 			}
-
 
 			if criterion.Operator == ":" {
 				token, err = p.expectOptionalWhitespaceFollowedBy(tokenizer, "right operand", assertlyExprMatcher, eof)