+++ testing/endpoint/http/service_contract.go
@@ -8,6 +8,7 @@ import (
 //ListenRequest represent HTTP endpoint listen request
 type ListenRequest struct {
 	Port          int
+	Rotate        bool
 	BaseDirectory string   `required:"true" description:"location with replay files (could be generate by https://github.com/viant/toolbox/blob/master/bridge/http_bridge_recording_util.go#L81"`
 	IndexKeys     []string `description:"recorded requests matching keys, by default: Method,URL,Body,Cookie,Content-Type"`
 }
@@ -34,6 +35,7 @@ func (r ListenRequest) AsHTTPServerTrips() *HTTPServerTrips {
 		r.IndexKeys = []string{MethodKey, URLKey, BodyKey, CookieKey, ContentTypeKey}
 	}
 	return &HTTPServerTrips{
+		Rotate:        r.Rotate,
 		BaseDirectory: r.BaseDirectory,
 		Trips:         make(map[string]*HTTPResponses),
 		IndexKeys:     r.IndexKeys,