+++ testing/runner/http/service.go
@@ -1,178 +1,105 @@
 package http
 
 import (
-	"bytes"
-	"encoding/base64"
 	"fmt"
 	"github.com/viant/endly"
+	"github.com/viant/endly/criteria"
 	"github.com/viant/endly/testing/validator"
 	"github.com/viant/endly/udf"
-	"github.com/viant/endly/util"
 	"github.com/viant/toolbox"
-	"github.com/viant/toolbox/data"
-	"io"
-	"io/ioutil"
 	"net/http"
-	"strings"
 	"time"
 )
 
 //ServiceID represents http runner service id.
 const ServiceID = "http/runner"
-
-//PreviousTripStateKey keys to store previous request details for multi trip HTTP Send request in context state
-const PreviousTripStateKey = "previous"
+const RunnerID = "HttpRunner"
 
 type service struct {
 	*endly.AbstractService
 }
 
-func (s *service) processResponse(context *endly.Context, sendRequest *SendRequest, sendHTTPRequest *Request, response *Response, httpResponse *http.Response, isBase64Encoded bool, extracted map[string]interface{}) (string, error) {
-	response.Header = make(map[string][]string)
-	copyHeaders(httpResponse.Header, response.Header)
-	readBody(httpResponse, response, isBase64Encoded)
-	if sendHTTPRequest.ResponseUdf != "" {
-		transformed, err := udf.TransformWithUDF(context, sendHTTPRequest.ResponseUdf, sendHTTPRequest.URL, response.Body)
-		if err != nil {
-			return "", err
-		}
-		if toolbox.IsMap(transformed) {
-			response.Body, _ = toolbox.AsJSONText(transformed)
-		} else {
-			response.Body = toolbox.AsString(transformed)
-		}
-	}
-
-	var responseBody = replaceResponseBodyIfNeeded(sendHTTPRequest, response.Body)
-	return responseBody, nil
-}
-
-func (s *service) sendRequest(context *endly.Context, client *http.Client, HTTPRequest *Request, sessionCookies *Cookies, sendGroupRequest *SendRequest, sendGroupResponse *SendResponse) error {
-	var err error
-	var state = context.State()
-	cookies := state.GetMap("cookies")
-	var reader io.Reader
-	var isBase64Encoded = false
-	HTTPRequest = HTTPRequest.Expand(context)
-	var body []byte
-	var ok bool
-	if len(HTTPRequest.Body) > 0 {
-		body = []byte(HTTPRequest.Body)
-		if HTTPRequest.RequestUdf != "" {
-			transformed, err := udf.TransformWithUDF(context, HTTPRequest.RequestUdf, HTTPRequest.URL, string(body))
-			if err != nil {
-				return err
-			}
-			if body, ok = transformed.([]byte); !ok {
-				body = []byte(toolbox.AsString(transformed))
-			}
-		}
-		isBase64Encoded = strings.HasPrefix(string(body), "base64:")
-		body, err = util.FromPayload(string(body))
-		if err != nil {
-			return err
-		}
-		reader = bytes.NewReader(body)
-	}
-
-	httpRequest, err := http.NewRequest(strings.ToUpper(HTTPRequest.Method), HTTPRequest.URL, reader)
+func (s *service) send(context *endly.Context, sendGroupRequest *SendRequest) (*SendResponse, error) {
+	client, err := toolbox.NewHttpClient(s.applyDefaultTimeoutIfNeeded(sendGroupRequest.Options)...)
 	if err != nil {
-		return err
+		return nil, fmt.Errorf("failed to send req: %v", err)
+	}
+	initializeContext(context)
+	defer s.resetContext(context, sendGroupRequest)
+
+	if err = udf.RegisterProviders(sendGroupRequest.UdfProviders); err != nil {
+		return nil, err
 	}
 
-	copyHeaders(HTTPRequest.Header, httpRequest.Header)
-	sessionCookies.SetHeader(HTTPRequest.Header)
-	HTTPRequest.Cookies.SetHeader(httpRequest.Header)
-
-	response := &Response{}
-	sendGroupResponse.Responses = append(sendGroupResponse.Responses, response)
-	startEvent := s.Begin(context, HTTPRequest)
-	repeater := HTTPRequest.Repeater.Init()
-
-	var HTTPResponse *http.Response
-	var responseBody string
-	var bodyCache []byte
-	var useCachedBody = repeater.Repeat > 1 && httpRequest.ContentLength > 0
-	if useCachedBody {
-		bodyCache, err = ioutil.ReadAll(httpRequest.Body)
-		if err != nil {
-			return err
-		}
+	var sendGroupResponse = &SendResponse{
+		Responses: make([]*Response, 0),
+		Data:      make(map[string]interface{}),
 	}
-
-	handler := func() (interface{}, error) {
-		if useCachedBody {
-			httpRequest.Body = ioutil.NopCloser(bytes.NewReader(bodyCache))
-		}
-
-		HTTPResponse, err = client.Do(httpRequest)
+	var sessionCookies Cookies = make([]*http.Cookie, 0)
+	for _, req := range sendGroupRequest.Requests {
+		err = s.sendRequest(context, client, req, &sessionCookies, sendGroupRequest, sendGroupResponse)
 		if err != nil {
 			return nil, err
 		}
-		responseBody, err = s.processResponse(context, sendGroupRequest, HTTPRequest, response, HTTPResponse, isBase64Encoded, sendGroupResponse.Data)
-		return responseBody, err
 	}
-
-	err = repeater.Run(s.AbstractService, "HTTPRunner", context, handler, sendGroupResponse.Data)
-	if err != nil {
-		return err
-	}
-	var responseCookies Cookies = HTTPResponse.Cookies()
-	response.Cookies = responseCookies.IndexByName()
-	for k, cookie := range response.Cookies {
-		cookies.Put(k, cookie.Value)
-	}
-	sessionCookies.AddCookies(responseCookies...)
-	endEvent := s.End(context)(startEvent, response)
-
-	var previous = state.GetMap(PreviousTripStateKey)
-	if previous == nil {
-		previous = data.NewMap()
-	}
-	response.Code = HTTPResponse.StatusCode
-	response.TimeTakenMs = int(startEvent.Timestamp().Sub(endEvent.Timestamp()) / time.Millisecond)
-
-	if toolbox.IsCompleteJSON(responseBody) {
-		response.JSONBody, err = toolbox.JSONToMap(responseBody)
-		if err == nil && HTTPRequest.Repeater != nil {
-			_ = HTTPRequest.Variables.Apply(data.Map(response.JSONBody), previous)
+	if sendGroupRequest.Expect != nil {
+		var actual = map[string]interface{}{
+			"Responses": sendGroupResponse.Responses,
+			"Data":      sendGroupResponse.Data,
 		}
+		sendGroupResponse.Assert, err = validator.Assert(context, sendGroupRequest, sendGroupRequest.Expect, actual, "HTTP.responses", "assert http responses")
 	}
+	return sendGroupResponse, nil
 
-	for k, v := range sendGroupResponse.Data {
-		var expanded = previous.Expand(v)
-		previous[k] = state.Expand(expanded)
-	}
-
-	err = repeater.Variables.Apply(previous, previous)
-	if err != nil {
-		return err
-	}
-	if len(previous) > 0 {
-		state.Put(PreviousTripStateKey, previous)
-	}
-	if HTTPRequest.When != "" {
-		return nil
-	}
-
-	for _, candidate := range sendGroupRequest.Requests {
-		if candidate.When != "" && strings.Contains(response.Body, candidate.When) {
-			err = s.sendRequest(context, client, candidate, sessionCookies, sendGroupRequest, sendGroupResponse)
-			if err != nil {
-				return err
-			}
-		}
-	}
-	return nil
 }
 
-func replaceResponseBodyIfNeeded(sendHTTPRequest *Request, responseBody string) string {
-	if len(sendHTTPRequest.Replace) > 0 {
-		for k, v := range sendHTTPRequest.Replace {
-			responseBody = strings.Replace(responseBody, k, v, len(responseBody))
-		}
+func (s *service) sendRequest(context *endly.Context, client *http.Client, request *Request, sessionCookies *Cookies, sendGroupRequest *SendRequest, sendGroupResponse *SendResponse) error {
+	var err error
+	var state = context.State()
+	cookies := state.GetMap("cookies")
+	trips := Trips(state.GetMap(TripsKey))
+	canRun, err := criteria.Evaluate(context, context.State(), request.When, fmt.Sprintf("%v.When", "HttpRequest"), true)
+	if err != nil || !canRun {
+		return err
 	}
-	return responseBody
+	httpRequest, expectBinary, err := request.Build(context, *sessionCookies)
+	if err != nil {
+		return err
+	}
+	trips.addRequest(request)
+	startEvent := s.Begin(context, request)
+	repeater := request.Repeater.Init()
+	var response *Response
+	bodyProvider, err := getRequestBodyReader(httpRequest, repeater.Repeat)
+
+	handler := func() (interface{}, error) {
+		httpRequest.Body = bodyProvider()
+		httpResponse, err := client.Do(httpRequest)
+		if err != nil {
+			return nil, err
+		}
+		if response == nil { //if request is repeated only the allocated one, and keep overriding it to see the last snapshot
+			response = sendGroupResponse.NewResponse()
+		}
+		response.Merge(httpResponse, expectBinary)
+		response.UpdateCookies(cookies)
+		sessionCookies.AddCookies(httpResponse.Cookies()...)
+		err = response.TransformBodyIfNeeded(context, request)
+		return response.Body, err
+	}
+
+	err = repeater.Run(s.AbstractService, RunnerID, context, handler, sendGroupResponse.Data)
+	if err != nil {
+		return err
+	}
+	if toolbox.IsCompleteJSON(response.Body) {
+		response.JSONBody, err = toolbox.JSONToMap(response.Body)
+	}
+	trips.setData(sendGroupResponse.Data)
+	trips.addResponse(response)
+	endEvent := s.End(context)(startEvent, response)
+	response.TimeTakenMs = int(endEvent.Timestamp().Sub(startEvent.Timestamp()) / time.Millisecond)
+	return nil
 }
 
 func (s *service) applyDefaultTimeoutIfNeeded(options []*toolbox.HttpOptions) []*toolbox.HttpOptions {
@@ -191,97 +118,9 @@ func (s *service) applyDefaultTimeoutIfNeeded(options []*toolbox.HttpOptions) []
 	}
 }
 
-func (s *service) send(context *endly.Context, request *SendRequest) (*SendResponse, error) {
-	client, err := toolbox.NewHttpClient(s.applyDefaultTimeoutIfNeeded(request.Options)...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to send req: %v", err)
-	}
-	defer s.resetContext(context, request)
-	var result = &SendResponse{
-		Responses: make([]*Response, 0),
-		Data:      make(map[string]interface{}),
-	}
-	var sessionCookies Cookies = make([]*http.Cookie, 0)
-	var state = context.State()
-	if !state.Has("cookies") {
-		state.Put("cookies", data.NewMap())
-	}
-	for _, req := range request.Requests {
-		if req.When != "" {
-			continue
-		}
-		err = s.sendRequest(context, client, req, &sessionCookies, request, result)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if request.Expect != nil {
-		var actual = map[string]interface{}{
-			"Responses": result.Responses,
-			"Data":      result.Data,
-		}
-		result.Assert, err = validator.Assert(context, request, request.Expect, actual, "HTTP.responses", "assert http responses")
-	}
-
-	return result, nil
-
-}
-func readBody(httpResponse *http.Response, response *Response, expectBased64Encoded bool) {
-
-	body, err := ioutil.ReadAll(httpResponse.Body)
-	if err != nil {
-		response.Error = fmt.Sprintf("%v", err)
-		return
-	}
-	httpResponse.Body.Close()
-	if expectBased64Encoded {
-		buf := new(bytes.Buffer)
-		encoder := base64.NewEncoder(base64.StdEncoding, buf)
-		_, _ = encoder.Write(body)
-		_ = encoder.Close()
-		response.Body = "base64:" + string(buf.Bytes())
-
-	} else {
-		response.Body = string(body)
-	}
-}
-
-func copyHeaders(source http.Header, target http.Header) {
-	for key, values := range source {
-		if _, has := target[key]; !has {
-			target[key] = make([]string, 0)
-		}
-		if len(values) == 1 {
-			target.Set(key, values[0])
-		} else {
-
-			for _, value := range values {
-				target.Add(key, value)
-			}
-		}
-	}
-}
-
-func copyExpandedHeaders(source http.Header, target http.Header, context *endly.Context) {
-	for key, values := range source {
-		if _, has := target[key]; !has {
-			target[key] = make([]string, 0)
-		}
-		if len(values) == 1 {
-			target.Set(key, context.Expand(values[0]))
-		} else {
-			for _, value := range values {
-				target.Add(key, context.Expand(value))
-			}
-		}
-	}
-}
-
 //resetContext resets context for variables with Reset flag set, and removes PreviousTripStateKey
 func (s *service) resetContext(context *endly.Context, request *SendRequest) {
 	state := context.State()
-	state.Delete(PreviousTripStateKey)
 	for _, request := range request.Requests {
 		if request.Repeater != nil && len(request.Extraction) > 0 {
 			request.Extraction.Reset(state)