+++ udf/udf_test.go
@@ -2,6 +2,8 @@ package udf
 
 import (
 	"github.com/stretchr/testify/assert"
+	"github.com/viant/assertly"
+	"github.com/viant/endly"
 	"github.com/viant/endly/test/proto"
 	"github.com/viant/toolbox"
 	"github.com/viant/toolbox/data"
@@ -34,7 +36,7 @@ func Test_AsProtobufMessage(t *testing.T) {
 	if assert.Nil(t, err) {
 		aMap := toolbox.AsMap(message)
 		assert.EqualValues(t, 1, aMap["Id"])
-		assert.EqualValues(t, "abc", aMap["Name"])
+		assert.EqualValues(t, "abc", aMap["Desc"])
 	}
 
 }
@@ -78,4 +80,51 @@ func TestURLPath(t *testing.T) {
 	var expr = `$URLPath($URL)`
 	var expanded = aMap.Expand(expr)
 	assert.EqualValues(t, "/abc", expanded)
+}
+
+var avroSchema = `{"type": "record", "name": "user", "fields": [{"name": "ID","type":"int"},{"name": "Desc","type":"string"}]}`
+
+func TestNewAvroWriterProvider(t *testing.T) {
+
+	writer, err := NewAvroWriterProvider(avroSchema, "snappy")
+	assert.Nil(t, err)
+	type user struct {
+		ID   int
+		Desc string
+	}
+	transformed, err := writer(&user{1, "test"}, nil)
+	assert.Nil(t, err)
+	assert.NotNil(t, transformed)
+	output, err := NewAvroReader(transformed, nil)
+	assert.Nil(t, err)
+	assertly.AssertValues(t, "{\"Desc\":\"test\",\"ID\":1}", output)
+}
+
+func TestRegisterProviders(t *testing.T) {
+	{
+		err := RegisterProviders([]*endly.UdfProvider{
+			{
+				Id:       "userAvro",
+				Provider: "AvroWriter",
+				Params: []interface{}{
+					avroSchema,
+				},
+			},
+		})
+		assert.Nil(t, err)
+		_, ok := endly.UdfRegistry["userAvro"]
+		assert.True(t, ok)
+	}
+	{
+		err := RegisterProviders([]*endly.UdfProvider{
+			{
+				Id:       "userAvro",
+				Provider: "AvroWritaaaaer",
+				Params: []interface{}{
+					avroSchema,
+				},
+			},
+		})
+		assert.NotNil(t, err)
+	}
 }